Block stray HTTP requests from the test suite
Makes "no test reaches the network" a property of the suite rather than
something each new test has to remember. An unfaked outbound request now
throws StrayRequestException; Http::fake() is how a test opts back in.

This is preventative, not remedial. The Discord tests were already faking
their webhook calls -- the 400 Invalid Form Body and 404 Unknown Webhook
lines in the dev log came from those stubs, not from Discord. Verified by
running the whole suite with the guard in place: 1501 tests, all green,
so nothing was relying on real network access.

Folds the mail guard added in the previous commit into the same test,
since it is one concern: the suite must not touch the outside world.
